requirements.txt
alembic==1.4.2
Babel==2.8.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
dnspython==1.16.0
dominate==2.5.1
elasticsearch==7.8.0
email-validator==1.1.1
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-Sendmail==0.1
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gmplot==1.4.1
googlemaps==4.4.1
guess-language==0.2
guess-language-spirit==0.5.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
WTForms-SQLAlchemy==0.1
